+++ src/services/alImagesService.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode';
-import { ALActionImageBrowser } from '../actionimagebrowser/alActionImageBrowser';
+import { ALImageBrowser } from '../imagebrowser/alImageBrowser';
 import { DevToolsExtensionContext } from "../devToolsExtensionContext";
 import { DevToolsExtensionService } from './devToolsExtensionService';
 
@@ -11,13 +11,18 @@ export class ALImagesService extends DevToolsExtensionService {
     }
 
     protected registerCommands() {
-        //al action images viewer
+        this.registerImageBrowserCommand('azALDevTools.viewActionImages', 'Action Images', 'actionImages', true, 'imageaction');
+        this.registerImageBrowserCommand('azALDevTools.viewCueGroupActionImages', 'CueGroup Action Images', 'actionCueGroupImages', true, 'imagecgaction');
+        this.registerImageBrowserCommand('azALDevTools.viewCueGroupFieldsImages', 'CueGroup Fields Images', 'fieldCueGroupImages', false, 'imagecgfld');
+        this.registerImageBrowserCommand('azALDevTools.viewRoleCenterActionImages', 'Role Center Action Images', 'roleCenterActionImages', true, 'imagercaction');
+    }
+
+    protected registerImageBrowserCommand(commandName: string, caption: string, imagesType: string, withActions: boolean, imageStyleType: string) {
         this._context.vscodeExtensionContext.subscriptions.push(
             vscode.commands.registerCommand(
-                'azALDevTools.viewActionImages',
-                () => {
-                    let actionImageBrowser: ALActionImageBrowser = new ALActionImageBrowser(this._context);
-                    actionImageBrowser.show();
+                commandName, () => {
+                    let imageBrowser: ALImageBrowser = new ALImageBrowser(this._context, caption, imagesType, withActions, imageStyleType);
+                    imageBrowser.show();
                 }
             ));
     }